Add process overview; Detail plus edge pieces
Add a brief overview of the process before digging into the minutiae of
the process.

Clarify the "plus" directions a bit.

Flesh out the details of the process for aligning the "plus"'s ends with
their respective sides.

Swap RIGHT and BACK in the diagrams, to make it a bit easier to follow
(makes them closer to unrolled cubes).